+++ src/listbee/types/__init__.py
@@ -29,7 +29,7 @@ from listbee.types.shared import (
     WebhookEventType,
     WebhookReadiness,
 )
-from listbee.types.signup import SignupResponse, VerifyResponse
+from listbee.types.signup import AuthSessionResponse, OtpRequestResponse
 from listbee.types.stripe import StripeConnectSessionResponse
 from listbee.types.utility import PingResponse
 from listbee.types.webhook import WebhookEventResponse, WebhookResponse, WebhookTestResponse
@@ -75,9 +75,9 @@ __all__ = [
     "CustomerResponse",
     # File
     "FileResponse",
-    # Signup
-    "SignupResponse",
-    "VerifyResponse",
+    # Auth/Signup
+    "OtpRequestResponse",
+    "AuthSessionResponse",
     # Stripe
     "StripeConnectSessionResponse",
     # Utility